Read sections in consistent order
Would read the sections in the order they appeared in before, which would cause some dependency problems.

e.g. Lots of sections require info from .names section but would appear before .names section in list. To fix this, .names was moved so that it is the first section to be read.